INTELLIJ-113 remove deprecated methods
