+++ src/main/java/com/liferay/ide/idea/core/LiferayPostStartupActivity.java
@@ -14,18 +14,31 @@
 
 package com.liferay.ide.idea.core;
 
+import com.intellij.ProjectTopics;
+import com.intellij.facet.Facet;
+import com.intellij.facet.FacetManager;
+import com.intellij.facet.FacetType;
+import com.intellij.facet.ModifiableFacetModel;
+import com.intellij.facet.ProjectFacetManager;
+import com.intellij.javaee.web.facet.WebFacet;
+import com.intellij.javaee.web.facet.WebFacetConfiguration;
+import com.intellij.javaee.web.facet.WebFacetType;
 import com.intellij.openapi.application.Application;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.externalSystem.service.project.manage.ProjectDataImportListener;
 import com.intellij.openapi.module.Module;
 import com.intellij.openapi.project.DumbAware;
+import com.intellij.openapi.project.ModuleListener;
 import com.intellij.openapi.project.Project;
+import com.intellij.openapi.roots.ModuleRootManager;
 import com.intellij.openapi.startup.StartupActivity;
 import com.intellij.openapi.startup.StartupManager;
+import com.intellij.openapi.vfs.LocalFileSystem;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.util.messages.MessageBus;
 import com.intellij.util.messages.MessageBusConnection;
 
+import com.liferay.ide.idea.util.FileUtil;
 import com.liferay.ide.idea.util.LiferayWorkspaceSupport;
 import com.liferay.ide.idea.util.ProjectConfigurationUtil;
 
@@ -38,7 +51,7 @@ import org.jetbrains.idea.maven.project.MavenProjectsManager;
 /**
  * @author Simon Jiang
  */
-public class LiferayPostStartupActivity implements DumbAware, LiferayWorkspaceSupport, StartupActivity {
+public class LiferayPostStartupActivity implements DumbAware, LiferayWorkspaceSupport, StartupActivity.Background {
 
 	@Override
 	public void runActivity(@NotNull Project project) {
@@ -72,6 +85,25 @@ public class LiferayPostStartupActivity implements DumbAware, LiferayWorkspaceSu
 						});
 				}));
 
+		startupManager.registerPostStartupActivity(
+			() -> messageBusConnection.subscribe(
+				ProjectTopics.MODULES,
+				new ModuleListener() {
+
+					@Override
+					public void moduleAdded(@NotNull Project project, @NotNull Module module) {
+						Application application = ApplicationManager.getApplication();
+
+						application.runWriteAction(
+							() -> {
+								if (LiferayWorkspaceSupport.isValidWorkspaceLocation(project)) {
+									_addWebRoot(module);
+								}
+							});
+					}
+
+				}));
+
 		messageBusConnection.subscribe(
 			MavenImportListener.TOPIC,
 			(projects, list) -> {
@@ -101,6 +133,71 @@ public class LiferayPostStartupActivity implements DumbAware, LiferayWorkspaceSu
 					}
 				);
 			});
+	}
+
+	private void _addWebRoot(Module module) {
+		ModuleRootManager moduleRootManager = ModuleRootManager.getInstance(module);
+
+		VirtualFile[] sourceRoots = moduleRootManager.getSourceRoots();
+
+		if (sourceRoots.length > 0) {
+			for (VirtualFile sourceRoot : sourceRoots) {
+				String sourcePath = sourceRoot.getPath();
+
+				if (sourcePath.contains("src/main/resources")) {
+					String resourcesPath = sourcePath.concat("/META-INF/resources");
+
+					LocalFileSystem localFileSystem = LocalFileSystem.getInstance();
+
+					VirtualFile resources = localFileSystem.findFileByPath(resourcesPath);
+
+					if (FileUtil.exist(resources)) {
+						boolean hasWebFacet = false;
+
+						FacetManager facetManager = FacetManager.getInstance(module);
+
+						Facet<?>[] facets = facetManager.getAllFacets();
+
+						for (Facet<?> facet : facets) {
+							WebFacetType webFacetType = WebFacetType.getInstance();
+
+							FacetType<?, ?> facetType = facet.getType();
+
+							String facetTypePresentableName = facetType.getPresentableName();
+
+							if (facetTypePresentableName.equals(webFacetType.getPresentableName())) {
+								hasWebFacet = true;
+
+								break;
+							}
+						}
+
+						if (!hasWebFacet) {
+							ProjectFacetManager projectFacetManager = ProjectFacetManager.getInstance(
+								module.getProject());
+
+							WebFacetConfiguration webFacetConfiguration =
+								projectFacetManager.createDefaultConfiguration(WebFacetType.getInstance());
+
+							ModifiableFacetModel modifiableFacetModel = facetManager.createModifiableModel();
+
+							WebFacetType webFacetType = WebFacetType.getInstance();
+
+							WebFacet webFacet = facetManager.createFacet(
+								webFacetType, webFacetType.getPresentableName(), webFacetConfiguration, null);
+
+							webFacet.addWebRoot(resources, "/");
+
+							modifiableFacetModel.addFacet(webFacet);
+
+							Application application = ApplicationManager.getApplication();
+
+							application.invokeLater(() -> application.runWriteAction(modifiableFacetModel::commit));
+						}
+					}
+				}
+			}
+		}
 	}
 
 }
