+++ src/main/resources/META-INF/javaee.xml
@@ -2,10 +2,4 @@
 	<extensions defaultExtensionNs="com.intellij">
 		<debugger.positionManagerFactory id="liferayPositionManager" implementation="com.liferay.ide.idea.language.jsp.LiferayJspDebuggerPositionManagerFactory" />
 	</extensions>
-
-	<project-components>
-		<component>
-			<implementation-class>com.liferay.ide.idea.project.LiferayProjectComponent</implementation-class>
-		</component>
-	</project-components>
 </idea-plugin>
