+++ src/test/java/com/liferay/ide/idea/language/service/LiferayServiceXMLLineMarkerProviderTest.java
@@ -20,7 +20,6 @@ import com.intellij.codeInsight.daemon.RelatedItemLineMarkerInfo;
 import com.intellij.navigation.GotoRelatedItem;
 import com.intellij.openapi.module.Module;
 import com.intellij.openapi.projectRoots.Sdk;
-import com.intellij.openapi.projectRoots.impl.JavaAwareProjectJdkTableImpl;
 import com.intellij.openapi.roots.ContentEntry;
 import com.intellij.openapi.roots.LanguageLevelModuleExtension;
 import com.intellij.openapi.roots.ModifiableRootModel;
@@ -29,6 +28,8 @@ import com.intellij.testFramework.IdeaTestUtil;
 import com.intellij.testFramework.LightProjectDescriptor;
 import com.intellij.testFramework.fixtures.DefaultLightProjectDescriptor;
 import com.intellij.testFramework.fixtures.LightJavaCodeInsightFixtureTestCase;
+
+import com.liferay.ide.idea.util.SdkUtil;
 
 import java.util.Collection;
 import java.util.List;
@@ -109,11 +110,7 @@ public class LiferayServiceXMLLineMarkerProviderTest extends LightJavaCodeInsigh
 				languageLevelModuleExtension.setLanguageLevel(LanguageLevel.JDK_1_8);
 			}
 
-			JavaAwareProjectJdkTableImpl javaAwareProjectJdkTableImpl = JavaAwareProjectJdkTableImpl.getInstanceEx();
-
-			Sdk sdk = javaAwareProjectJdkTableImpl.getInternalJdk();
-
-			modifiableRootModel.setSdk(sdk);
+			modifiableRootModel.setSdk(SdkUtil.getTestJdk());
 		}
 
 		@Override
