+++ src/test/java/com/liferay/ide/idea/language/tag/LiferayTaglibStrictQuoteEscapingInspectionTest.java
@@ -17,7 +17,6 @@ package com.liferay.ide.idea.language.tag;
 import com.intellij.codeInsight.intention.IntentionAction;
 import com.intellij.openapi.module.Module;
 import com.intellij.openapi.projectRoots.Sdk;
-import com.intellij.openapi.projectRoots.impl.JavaAwareProjectJdkTableImpl;
 import com.intellij.openapi.roots.ContentEntry;
 import com.intellij.openapi.roots.LanguageLevelModuleExtension;
 import com.intellij.openapi.roots.ModifiableRootModel;
@@ -26,6 +25,8 @@ import com.intellij.testFramework.IdeaTestUtil;
 import com.intellij.testFramework.LightProjectDescriptor;
 import com.intellij.testFramework.fixtures.DefaultLightProjectDescriptor;
 import com.intellij.testFramework.fixtures.LightJavaCodeInsightFixtureTestCase;
+
+import com.liferay.ide.idea.util.SdkUtil;
 
 import java.util.List;
 import java.util.Objects;
@@ -89,11 +90,7 @@ public class LiferayTaglibStrictQuoteEscapingInspectionTest extends LightJavaCod
 				languageLevelModuleExtension.setLanguageLevel(LanguageLevel.JDK_1_8);
 			}
 
-			JavaAwareProjectJdkTableImpl javaAwareProjectJdkTableImpl = JavaAwareProjectJdkTableImpl.getInstanceEx();
-
-			Sdk jdk = javaAwareProjectJdkTableImpl.getInternalJdk();
-
-			modifiableRootModel.setSdk(jdk);
+			modifiableRootModel.setSdk(SdkUtil.getTestJdk());
 		}
 
 		@Override
